Use POST method for logout and added CSRF protection && refactoring show page to add CSRF protection on delete as well

+++ src/Controller/PinsController.php
@@ -86,15 +86,18 @@ class PinsController extends AbstractController
     }
 
     /**
-     * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods="GET")
+     * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods="POST")
      */
     public function delete(Request $request, EntityManagerInterface $em, Pin $pin): Response
     {
-        $em->remove($pin);
-        $em->flush();
+        if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('_csrf_token'))) {
+            $em->remove($pin);
+            $em->flush();
 
-        $this->addFlash('info', 'Pin successfully deleted!');
-
+            $this->addFlash('info', 'Pin successfully deleted!');
+        } else {
+            $this->addFlash('error', 'Invalid Csrf token!');
+        }
 
         return $this->redirectToRoute('app_home');
     }
